Adding GitHub action for winget-releaser
- This is meant to automate releases to winget-pkgs
- Resolves #848 and should work as soon as microsoft/winget-pkgs#184453
  has been merged (baseline manifest for the 2.2.0 version)
- DO NOT MERGE BEFORE PREREQUISITES ARE IN PLACE (see below)

## Instructions for prerequisites
The `WINGET_TOKEN` needs to be created and registered in the repo _which
does the release_. For good measure we need to clone `winget-pkgs` on the
same user (or org) or set via `fork-user` it as per the README for
[winget-releaser](https://github.com/vedantmgoyal9/winget-releaser).

Configure the [Pull app](https://github.com/apps/pull) to keep the
`winget-pkgs` fork in sync with its upstream.

That `winget-pkgs` fork is going to be the _source_ of the PR filed at
`microsoft/winget-pkgs` and the PRs are being opened on behalf of the
owner of the used token.

1. Head to https://github.com/settings/tokens and create a new (classic)
   personal access token with _only_ the `public_repo` scope activated for
   it
2. In https://github.com/OWASP/threat-dragon/settings/secrets/actions
   create a secret named WINGET_TOKEN (as per vedantmgoyal9/winget-releaser)
3. Merge this PR
4. Create a (subsequent) release